+++ vitest.config.ts
@@ -1,10 +1,8 @@
-import type { UserConfig } from "vitest/config";
+import { defineConfig } from "vitest/config";
 
-const config: UserConfig = {
+export default defineConfig({
   test: {
     environment: "node",
     pool: "forks",
   },
-};
-
-export default config;
+});